requirements.txt
pandas==3.0.0
python-dotenv==1.2.1
pydantic==2.13.0b2
tenacity==9.1.2
openai==2.26.0
google-genai==1.66.0
ollama==0.6.1
# Langchain Core
langchain==1.2.12
langchain-core==1.2.19
# Langchain Provider Integrations
langchain-openai==1.1.11
langchain-google-genai==4.2.1
langchain-community==0.4.1

